requirements.txt
opencv-python
websockets
face_recognition
dlib
RPi.GPIO